move rsvp component to shared folder
