+++ src/pages/rsvp.js
@@ -4,7 +4,7 @@ import Layout from 'components/layout';
 import RsvpHeader from 'components/rsvp/header';
 import DesktopNav from 'components/shared/desktopNav';
 import MobileNav from 'components/shared/mobileNav';
-import Rsvp from 'components/home/rsvp';
+import Rsvp from 'components/shared/rsvp';
 
 const RsvpPage = () => (
   <Layout>
